Move backendMetadataFetch call to the place that is used and remove ApmBackendContext
